feat: prefix project routes with admin- to improve route namespacing

+++ routes/web.php
@@ -72,12 +72,12 @@ Route::middleware([
 
 
     
-    Route::get('projects', [ProjectController::class, 'index'])->name('projects.index');
-    Route::get('projects/create', [ProjectController::class, 'create'])->name('projects.create');
-    Route::post('projects', [ProjectController::class, 'store'])->name('projects.store');
-    Route::get('projects/{project}/edit', [ProjectController::class, 'edit'])->name('projects.edit');
-    Route::put('projects/{project}', [ProjectController::class, 'update'])->name('projects.update');
-    Route::delete('projects/{project}', [ProjectController::class, 'destroy'])->name('projects.destroy');
+    Route::get('admin-projects', [ProjectController::class, 'index'])->name('projects.index');
+    Route::get('admin-projects/create', [ProjectController::class, 'create'])->name('projects.create');
+    Route::post('admin-projects', [ProjectController::class, 'store'])->name('projects.store');
+    Route::get('admin-projects/{project}/edit', [ProjectController::class, 'edit'])->name('projects.edit');
+    Route::put('admin-projects/{project}', [ProjectController::class, 'update'])->name('projects.update');
+    Route::delete('admin-projects/{project}', [ProjectController::class, 'destroy'])->name('projects.destroy');
 
 
     Route::get('news',[NewsController::class, 'index']);
